Calculate trucks
Truck trucks trucks

+++ src/resource_calculation.h
@@ -1,4 +1,4 @@
 #include <package_distribution.h>
 
 package_t temp_package();
-int calculate_trucks(int orders, package_t package);
+double calculate_trucks(int orders, package_t package);
